requirements.txt
Click==7.0
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
djangorestframework==3.9.2
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mock==2.0.0
numpy==1.16.0
pbr==5.1.0
psycopg2==2.7.7
pytz==2018.9
selenium==3.14.1
six==1.11.0
urllib3==1.24
virtualenv==16.3.0
Werkzeug==0.14.1
whitenoise==4.1.2